Change Mock Entity Clean Code

+++ presentation/src/main/kotlin/app/presentation/base/adapter/RecyclerItem.kt
@@ -1,7 +1,7 @@
 package app.presentation.base.adapter
 
 interface RecyclerItem {
-    val id: Int?
+    val id: Long?
     override fun equals(other: Any?): Boolean
     override fun hashCode(): Int
 }
